+++ functions/public/New-ModuleFromFiles.ps1
@@ -1,4 +1,3 @@
-
 Function New-ModuleFromFiles {
     [cmdletbinding(SupportsShouldProcess)]
     Param(
@@ -16,7 +15,7 @@ Function New-ModuleFromFiles {
             HelpMessage = "What is the parent path for the new module? It must already exist."
         )]
         [ValidateNotNullOrEmpty()]
-        [ValidateScript( { Test-Path $_ })]
+        [ValidateScript({Test-Path $_ })]
         [string]$ParentPath,
 
         [Parameter(
@@ -30,7 +29,7 @@ Function New-ModuleFromFiles {
             HelpMessage = "Enter the paths to PowerShell script files with functions to export."
         )]
         [ValidateNotNullOrEmpty()]
-        [ValidateScript( { Test-Path $_ })]
+        [ValidateScript({Test-Path $_ })]
         [string[]]$Files,
 
         [Parameter(
@@ -136,7 +135,7 @@ Function New-ModuleFromFiles {
         #create the root module
         $psm1 = @"
 
-Get-Childitem `$psscriptroot\functions\*.ps1 -recurse |
+Get-Childitem `$psscriptroot\$functionPath\*.ps1 -recurse |
 Foreach-Object {
 . `$_.FullName
 }
@@ -194,30 +193,3 @@ Foreach-Object {
     }
 } #end function
 
-#helper functions
-
-Function _mkHelp {
-    [cmdletbinding()]
-    Param(
-        #the path to the psd1 file
-        [string]$ModulePath,
-        #where to put the md files
-        [string]$Markdownpath,
-        #where to put the xml output
-        [string]$OutputPath
-    )
-
-    Write-Verbose "Invoking Import-Module on $modulepath"
-    Import-Module -Name $ModulePath -scope global
-    $ModuleName = (Get-Item $ModulePath).BaseName
-    Get-Command -Module $NewModuleName -OutVariable modcmds | Out-String | Write-Verbose
-    Write-Verbose "Invoking New-MarkdownHelp for module $modulename"
-    Try {
-        New-MarkdownHelp -Module $modulename -OutputFolder $Markdownpath -Force -ErrorAction Stop
-        Write-Verbose "Invoking New-Externalhelp to $outputPath"
-        New-ExternalHelp -Path $Markdownpath -OutputPath $OutputPath -Force -ErrorAction Stop
-    }
-    Catch {
-        Write-Warning "Failed to generate help content. $($_.exception.message)"
-    }
-}